TOOL: renames internal folder from counter to countable-tree
